requirements.txt
aiofiles==0.6.0
anyconfig==0.10.1
arrow==1.1.1
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.3
cryptography==3.4.8
decorator==5.0.9
dynaconf==3.1.5
ecdsa==0.17.0
fastapi==0.66.1
fsspec==0.8.7
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.1
h11==0.12.0
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
jmespath==0.10.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kedro==0.17.4
kedro-viz==3.14.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
numpy==1.21.2
pandas==1.3.2
parso==0.8.2
passlib==1.7.4
pickleshare==0.7.5
pip-tools==5.5.0
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
python-dateutil==2.8.2
python-jose==3.3.0
python-json-logger==2.0.2
python-multipart==0.0.5
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.26.0
rsa==4.7.2
semver==2.13.0
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.23
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
toposort==1.6
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.13.4
watchgod==0.7
wcwidth==0.2.5
wincertstore==0.2